+++ lib/src/models/ios_carrier_data.dart
@@ -6,16 +6,29 @@ class IosCarrierData {
   final List<CarrierData> carrierData;
   final bool supportsEmbeddedSIM;
   final List<String> carrierRadioAccessTechnologyTypeList;
+  final bool isSIMInserted;
+  final SubscriberInfo? subscriberInfo;
+  final CellularPlanInfo? cellularPlanInfo;
+  final NetworkStatus? networkStatus;
+
   IosCarrierData({
     required this.carrierData,
     required this.supportsEmbeddedSIM,
     required this.carrierRadioAccessTechnologyTypeList,
+    required this.isSIMInserted,
+    this.subscriberInfo,
+    this.cellularPlanInfo,
+    this.networkStatus,
   });
 
   IosCarrierData copyWith({
     List<CarrierData>? carrierData,
     bool? supportsEmbeddedSIM,
     List<String>? carrierRadioAccessTechnologyTypeList,
+    bool? isSIMInserted,
+    SubscriberInfo? subscriberInfo,
+    CellularPlanInfo? cellularPlanInfo,
+    NetworkStatus? networkStatus,
   }) {
     return IosCarrierData(
       carrierData: carrierData ?? this.carrierData,
@@ -23,6 +36,10 @@ class IosCarrierData {
       carrierRadioAccessTechnologyTypeList:
           carrierRadioAccessTechnologyTypeList ??
               this.carrierRadioAccessTechnologyTypeList,
+      isSIMInserted: isSIMInserted ?? this.isSIMInserted,
+      subscriberInfo: subscriberInfo ?? this.subscriberInfo,
+      cellularPlanInfo: cellularPlanInfo ?? this.cellularPlanInfo,
+      networkStatus: networkStatus ?? this.networkStatus,
     );
   }
 
@@ -32,6 +49,10 @@ class IosCarrierData {
       'supportsEmbeddedSIM': supportsEmbeddedSIM,
       'carrierRadioAccessTechnologyTypeList':
           carrierRadioAccessTechnologyTypeList,
+      'isSIMInserted': isSIMInserted,
+      'subscriberInfo': subscriberInfo?.toMap(),
+      'cellularPlanInfo': cellularPlanInfo?.toMap(),
+      'networkStatus': networkStatus?.toMap(),
     };
   }
 
@@ -39,12 +60,23 @@ class IosCarrierData {
     return IosCarrierData(
       carrierData: List<CarrierData>.from(
         map['carrierData']?.map(
-          (x) => CarrierData.fromMap(x),
-        ),
+              (x) => CarrierData.fromMap(x),
+            ) ??
+            [],
       ),
       supportsEmbeddedSIM: map['supportsEmbeddedSIM'] ?? false,
       carrierRadioAccessTechnologyTypeList:
-          List<String>.from(map['carrierRadioAccessTechnologyTypeList']),
+          List<String>.from(map['carrierRadioAccessTechnologyTypeList'] ?? []),
+      isSIMInserted: map['isSIMInserted'] ?? false,
+      subscriberInfo: map['subscriberInfo'] != null
+          ? SubscriberInfo.fromMap(map['subscriberInfo'])
+          : null,
+      cellularPlanInfo: map['cellularPlanInfo'] != null
+          ? CellularPlanInfo.fromMap(map['cellularPlanInfo'])
+          : null,
+      networkStatus: map['networkStatus'] != null
+          ? NetworkStatus.fromMap(map['networkStatus'])
+          : null,
     );
   }
 
@@ -55,7 +87,7 @@ class IosCarrierData {
 
   @override
   String toString() {
-    return 'IosCarrierData(carrierData: $carrierData, supportsEmbeddedSIM: $supportsEmbeddedSIM, carrierRadioAccessTechnologyTypeList: $carrierRadioAccessTechnologyTypeList)';
+    return 'IosCarrierData(carrierData: $carrierData, supportsEmbeddedSIM: $supportsEmbeddedSIM, carrierRadioAccessTechnologyTypeList: $carrierRadioAccessTechnologyTypeList, isSIMInserted: $isSIMInserted, subscriberInfo: $subscriberInfo, cellularPlanInfo: $cellularPlanInfo, networkStatus: $networkStatus)';
   }
 
   @override
@@ -66,29 +98,38 @@ class IosCarrierData {
         listEquals(other.carrierData, carrierData) &&
         other.supportsEmbeddedSIM == supportsEmbeddedSIM &&
         listEquals(other.carrierRadioAccessTechnologyTypeList,
-            carrierRadioAccessTechnologyTypeList);
+            carrierRadioAccessTechnologyTypeList) &&
+        other.isSIMInserted == isSIMInserted &&
+        other.subscriberInfo == subscriberInfo &&
+        other.cellularPlanInfo == cellularPlanInfo &&
+        other.networkStatus == networkStatus;
   }
 
   @override
   int get hashCode {
     return carrierData.hashCode ^
         supportsEmbeddedSIM.hashCode ^
-        carrierRadioAccessTechnologyTypeList.hashCode;
+        carrierRadioAccessTechnologyTypeList.hashCode ^
+        isSIMInserted.hashCode ^
+        subscriberInfo.hashCode ^
+        cellularPlanInfo.hashCode ^
+        networkStatus.hashCode;
   }
 }
 
 class CarrierData {
-  final String mobileNetworkCode;
-  final bool carrierAllowsVOIP;
-  final String mobileCountryCode;
-  final String carrierName;
-  final String isoCountryCode;
+  final String? mobileNetworkCode;
+  final bool? carrierAllowsVOIP;
+  final String? mobileCountryCode;
+  final String? carrierName;
+  final String? isoCountryCode;
+
   CarrierData({
-    required this.mobileNetworkCode,
-    required this.carrierAllowsVOIP,
-    required this.mobileCountryCode,
-    required this.carrierName,
-    required this.isoCountryCode,
+    this.mobileNetworkCode,
+    this.carrierAllowsVOIP,
+    this.mobileCountryCode,
+    this.carrierName,
+    this.isoCountryCode,
   });
 
   CarrierData copyWith({
@@ -119,11 +160,11 @@ class CarrierData {
 
   factory CarrierData.fromMap(Map<dynamic, dynamic> map) {
     return CarrierData(
-      mobileNetworkCode: map['mobileNetworkCode'] ?? '',
-      carrierAllowsVOIP: map['carrierAllowsVOIP'] ?? false,
-      mobileCountryCode: map['mobileCountryCode'] ?? '',
-      carrierName: map['carrierName'] ?? '',
-      isoCountryCode: map['isoCountryCode'] ?? '',
+      mobileNetworkCode: map['mobileNetworkCode'],
+      carrierAllowsVOIP: map['carrierAllowsVOIP'],
+      mobileCountryCode: map['mobileCountryCode'],
+      carrierName: map['carrierName'],
+      isoCountryCode: map['isoCountryCode'],
     );
   }
 
@@ -156,5 +197,196 @@ class CarrierData {
         mobileCountryCode.hashCode ^
         carrierName.hashCode ^
         isoCountryCode.hashCode;
+  }
+}
+
+class SubscriberInfo {
+  final int subscriberCount;
+  final List<String> subscriberIdentifiers;
+  final List<String>? carrierTokens;
+
+  SubscriberInfo({
+    required this.subscriberCount,
+    required this.subscriberIdentifiers,
+    this.carrierTokens,
+  });
+
+  SubscriberInfo copyWith({
+    int? subscriberCount,
+    List<String>? subscriberIdentifiers,
+    List<String>? carrierTokens,
+  }) {
+    return SubscriberInfo(
+      subscriberCount: subscriberCount ?? this.subscriberCount,
+      subscriberIdentifiers:
+          subscriberIdentifiers ?? this.subscriberIdentifiers,
+      carrierTokens: carrierTokens ?? this.carrierTokens,
+    );
+  }
+
+  Map<dynamic, dynamic> toMap() {
+    return {
+      'subscriberCount': subscriberCount,
+      'subscriberIdentifiers': subscriberIdentifiers,
+      'carrierTokens': carrierTokens,
+    };
+  }
+
+  factory SubscriberInfo.fromMap(Map<dynamic, dynamic> map) {
+    return SubscriberInfo(
+      subscriberCount: map['subscriberCount'] ?? 0,
+      subscriberIdentifiers:
+          List<String>.from(map['subscriberIdentifiers'] ?? []),
+      carrierTokens: map['carrierTokens'] != null
+          ? List<String>.from(map['carrierTokens'])
+          : null,
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory SubscriberInfo.fromJson(String source) =>
+      SubscriberInfo.fromMap(json.decode(source));
+
+  @override
+  String toString() {
+    return 'SubscriberInfo(subscriberCount: $subscriberCount, subscriberIdentifiers: $subscriberIdentifiers, carrierTokens: $carrierTokens)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is SubscriberInfo &&
+        other.subscriberCount == subscriberCount &&
+        listEquals(other.subscriberIdentifiers, subscriberIdentifiers) &&
+        listEquals(other.carrierTokens, carrierTokens);
+  }
+
+  @override
+  int get hashCode {
+    return subscriberCount.hashCode ^
+        subscriberIdentifiers.hashCode ^
+        carrierTokens.hashCode;
+  }
+}
+
+class CellularPlanInfo {
+  final bool supportsEmbeddedSIM;
+
+  CellularPlanInfo({
+    required this.supportsEmbeddedSIM,
+  });
+
+  CellularPlanInfo copyWith({
+    bool? supportsEmbeddedSIM,
+  }) {
+    return CellularPlanInfo(
+      supportsEmbeddedSIM: supportsEmbeddedSIM ?? this.supportsEmbeddedSIM,
+    );
+  }
+
+  Map<dynamic, dynamic> toMap() {
+    return {
+      'supportsEmbeddedSIM': supportsEmbeddedSIM,
+    };
+  }
+
+  factory CellularPlanInfo.fromMap(Map<dynamic, dynamic> map) {
+    return CellularPlanInfo(
+      supportsEmbeddedSIM: map['supportsEmbeddedSIM'] ?? false,
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory CellularPlanInfo.fromJson(String source) =>
+      CellularPlanInfo.fromMap(json.decode(source));
+
+  @override
+  String toString() {
+    return 'CellularPlanInfo(supportsEmbeddedSIM: $supportsEmbeddedSIM)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is CellularPlanInfo &&
+        other.supportsEmbeddedSIM == supportsEmbeddedSIM;
+  }
+
+  @override
+  int get hashCode {
+    return supportsEmbeddedSIM.hashCode;
+  }
+}
+
+class NetworkStatus {
+  final bool hasCellularData;
+  final int? activeServices;
+  final List<String>? technologies;
+
+  NetworkStatus({
+    required this.hasCellularData,
+    this.activeServices,
+    this.technologies,
+  });
+
+  NetworkStatus copyWith({
+    bool? hasCellularData,
+    int? activeServices,
+    List<String>? technologies,
+  }) {
+    return NetworkStatus(
+      hasCellularData: hasCellularData ?? this.hasCellularData,
+      activeServices: activeServices ?? this.activeServices,
+      technologies: technologies ?? this.technologies,
+    );
+  }
+
+  Map<dynamic, dynamic> toMap() {
+    return {
+      'hasCellularData': hasCellularData,
+      'activeServices': activeServices,
+      'technologies': technologies,
+    };
+  }
+
+  factory NetworkStatus.fromMap(Map<dynamic, dynamic> map) {
+    return NetworkStatus(
+      hasCellularData: map['hasCellularData'] ?? false,
+      activeServices: map['activeServices'],
+      technologies: map['technologies'] != null
+          ? List<String>.from(map['technologies'])
+          : null,
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory NetworkStatus.fromJson(String source) =>
+      NetworkStatus.fromMap(json.decode(source));
+
+  @override
+  String toString() {
+    return 'NetworkStatus(hasCellularData: $hasCellularData, activeServices: $activeServices, technologies: $technologies)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is NetworkStatus &&
+        other.hasCellularData == hasCellularData &&
+        other.activeServices == activeServices &&
+        listEquals(other.technologies, technologies);
+  }
+
+  @override
+  int get hashCode {
+    return hasCellularData.hashCode ^
+        activeServices.hashCode ^
+        technologies.hashCode;
   }
 }